# Conflicts:
#	src/main/java/dk/kea/webshop/controller/HomeController.java
#	src/main/java/dk/kea/webshop/model/Product.java
#	src/main/java/dk/kea/webshop/repository/ProductRepository.java

+++ src/main/java/dk/kea/webshop/service/ProductService.java
@@ -33,11 +33,11 @@ public class ProductService {
         return updateOK;
     }
 
-    public Product read(long id){
+    public Product read(int id){
         return productRepository.read(id);
     }
 
-    public boolean delete(Long id){
+    public boolean delete(int id){
         return productRepository.delete(id);
     }
